+++ core/lib/circuit/src/witness/noop.rs
@@ -59,7 +59,7 @@ pub fn noop_operation(tree: &CircuitAccountTree, acc_id: u32) -> Operation<Bn256
             pub_nonce: Some(Fr::zero()),
             new_pub_key_hash: Some(Fr::zero()),
             valid_from: Some(Fr::zero()),
-            valid_until: Some(Fr::from_str(&u64::MAX.to_string()).unwrap()),
+            valid_until: Some(Fr::from_str(&u32::MAX.to_string()).unwrap()),
         },
         lhs: OperationBranch {
             address: Some(account_address_fe),
